Changed hash type from uint32 to uint64. Fixed commit constants buffer to device in D3D11RHI. Fixed containing FDrawingPolicyLink in std::set
git-svn-id: svn://192.168.0.101/eleot@110 7f037352-864a-4b62-bc5a-4f07d5e00ced

+++ Develompent/Source/Engine/Core/Include/Core.h
@@ -32,7 +32,7 @@ enum ENoInit
  * @ingroup Core
  * Invalid hash
  */
-enum { INVALID_HASH = ( uint32 )-1 };
+enum { INVALID_HASH = ( uint64 )-1 };
 
 /**
  * @ingroup Core
